requirements.txt
Django==5.2.7
psycopg2-binary==2.9.9
djangorestframework==3.15.2
python-telegram-bot==21.6
whitenoise==6.7.0
dj-database-url==2.1.0
requests==2.32.3
django-cors-headers==4.4.0
